StudyBuddyAi/requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
langchain-community==0.2.16
python-dotenv==1.0.1
streamlit==1.37.1

pydantic==2.8.2
tenacity==8.5.0